# iuroadmap.services/api-gateway/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "User login (proxies to Auth service)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/learner": {
      "post": {
        "operationId": "AuthController_registerLearner",
        "summary": "Register a learner student account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Learner registered successfully"
          },
          "400": {
            "description": "Bad Request / Validation Error"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/mentor": {
      "post": {
        "operationId": "AuthController_registerMentor",
        "summary": "Register a mentor account along with profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MentorRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mentor registered successfully"
          },
          "400": {
            "description": "Bad Request / Validation Error"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout current user",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset token/link",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email requested"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password using token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "Get current authenticated user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/departments": {
      "post": {
        "operationId": "DepartmentsController_createDepartment",
        "summary": "Create a new department (Admin only)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Departments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DepartmentsController_viewDepartmentList",
        "summary": "Get list of all departments (Admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Departments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Departments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/departments/{id}": {
      "get": {
        "operationId": "DepartmentsController_viewDepartmentById",
        "summary": "Get department details by ID (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Department ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Department not found"
          }
        },
        "tags": [
          "Departments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "DepartmentsController_updateDepartment",
        "summary": "Update an existing department (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Department ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Departments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "DepartmentsController_deleteDepartment",
        "summary": "Delete a department by ID (Admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Department ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Department deleted successfully"
          }
        },
        "tags": [
          "Departments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/majors": {
      "get": {
        "operationId": "MajorsController_listMajors",
        "summary": "List all major roadmaps (Admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of major roadmaps retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MajorResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (Admin required)"
          }
        },
        "tags": [
          "Admin-Majors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/majors/{slug}": {
      "get": {
        "operationId": "MajorsController_getMajorBySlug",
        "summary": "Get a major roadmap by slug (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Major slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Major roadmap retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MajorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Major with slug not found"
          }
        },
        "tags": [
          "Admin-Majors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "MajorsController_updateMajor",
        "summary": "Update major metadata by slug (Admin only)",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Major slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMajorMetaPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Major metadata updated successfully"
          }
        },
        "tags": [
          "Admin-Majors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IUROADMAP API Gateway",
    "description": "The API Gateway for the IUROADMAP microservices network.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the user",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user (minimum 6 characters)",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthLoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT Access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
      "LearnerRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the learner",
            "example": "learner@example.com"
          },
          "password": {
            "type": "string",
            "description": "The account password (minimum 6 characters)",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "description": "Full name of the learner",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": [
              "STUDENT"
            ],
            "example": "STUDENT"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MentorRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the mentor",
            "example": "mentor@example.com"
          },
          "password": {
            "type": "string",
            "description": "The account password (minimum 6 characters)",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "description": "Full name of the mentor",
            "example": "Dr. Jane Smith"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": [
              "MENTOR"
            ],
            "example": "MENTOR"
          },
          "bio": {
            "type": "string",
            "description": "A brief biography or summary of experience",
            "example": "Senior Data Scientist with 10+ years in Fintech."
          },
          "cvUrl": {
            "type": "string",
            "description": "URL to the mentor CV/Resume",
            "example": "https://example.com/cv/janesmith.pdf"
          },
          "linkedinUrl": {
            "type": "string",
            "description": "LinkedIn profile URL",
            "example": "https://linkedin.com/in/janesmith"
          },
          "industry": {
            "type": "string",
            "description": "Industry category",
            "example": "BUSINESS_ANALYTICS"
          },
          "skills": {
            "description": "Array of skill tags",
            "example": [
              "SQL",
              "Python",
              "Machine Learning"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role",
          "cvUrl",
          "industry",
          "skills"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the user requesting password reset",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The password reset token sent via email",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "newPassword": {
            "type": "string",
            "description": "The new password (minimum 6 characters)",
            "example": "newStrongPassword123"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Department slug",
            "example": "school-of-computing"
          },
          "name": {
            "type": "string",
            "description": "Department name",
            "example": "School of Computing and Information Technology"
          },
          "description": {
            "type": "string",
            "description": "Department description",
            "example": "Leading research and education in computing"
          }
        },
        "required": [
          "slug",
          "name"
        ]
      },
      "DepartmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Department ID",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Department slug",
            "example": "school-of-computing"
          },
          "name": {
            "type": "string",
            "description": "Department name",
            "example": "School of Computing and Information Technology"
          },
          "description": {
            "type": "string",
            "description": "Department description",
            "example": "Leading research and education in computing"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2026-07-08T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp",
            "example": "2026-07-08T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Department slug",
            "example": "school-of-computing"
          },
          "name": {
            "type": "string",
            "description": "Department name",
            "example": "School of Computing and Information Technology"
          },
          "description": {
            "type": "string",
            "description": "Department description",
            "example": "Updated department description"
          }
        }
      },
...
      "MajorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Major ID",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Major slug",
            "example": "software-engineering"
          },
          "name": {
            "type": "string",
            "description": "Major name",
            "example": "Software Engineering"
          },
          "description": {
            "type": "string",
            "description": "Major description",
            "example": "Comprehensive SE roadmap"
          },
          "totalCreditsRequired": {
            "type": "number",
            "description": "Total credits required",
            "example": 130
          },
          "totalCourses": {
            "type": "number",
            "description": "Total courses included",
            "example": 42
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "totalCreditsRequired",
          "totalCourses"
        ]
      },
      "UpdateMajorMetaPayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Major name",
            "example": "Software Engineering"
          },
          "description": {
            "type": "string",
            "description": "Major description",
            "example": "Updated curriculum description"
          },
          "totalCreditsRequired": {
            "type": "number",
            "description": "Total credits required",
            "example": 130
          }
        }
      },
...
    }
  }
}